active_merchant: first pass at optimizing dispatch_to_gateways
Lazily retrieve the Kill Bill transaction and the last plugin transaction.

The order_id passed to the gateway is now the Kill Bill payment transaction id. To
set it to the payment transaction external key, pass external_key_as_order_id=true
in the plugin properties.

This is part of https://github.com/killbill/killbill-plugin-framework-ruby/issues/31.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>